feat: harden Git fetch and cache operations for WDL modules
Added the `[modules] max_transfer_bytes` cap so a runaway fetch aborts
instead of filling the module cache, fixed `ssh-agent` authentication by
returning only the credential types `libgit2` requested, and taught cache
reuse to evict a leaf whose sparse-checkout metadata cannot be parsed.

Split `resolver::git::ops` into `cache_store`, `checkout`, `creds`,
`error`, and `remote` submodules, since the single file had grown past
1,600 lines.
