+++ crates/wdl-modules/CHANGELOG.md
@@ -24,6 +24,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   remotes and local paths to URLs, and `git_remote_kind` for classifying a
   remote string
   ([#999](https://github.com/stjude-rust-labs/sprocket/pull/999)).
+* Added `ModulesConfig::max_transfer_bytes` along with the `TransferLimit` and
+  `TransferLimitError` types, which cap the bytes accepted from a Git remote
+  during one fetch
+  ([#1115](https://github.com/stjude-rust-labs/sprocket/pull/1115)).
 
 #### Changed
 
@@ -49,6 +53,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * `ManifestError::InvalidDependencyName` is now a struct variant carrying the
   offending `name` and the underlying `DependencyNameError` as its source
   ([#999](https://github.com/stjude-rust-labs/sprocket/pull/999)).
+
+#### Fixed
+
+* Git remotes reached over SSH now authenticate through `ssh-agent`; the
+  credential callback returns only credential types that `libgit2` requested
+  ([#1115](https://github.com/stjude-rust-labs/sprocket/pull/1115)).
+* A failing Git credential helper now produces a `git2` error with
+  `ErrorCode::Auth`, so the resolver reports it as an authentication failure
+  ([#1115](https://github.com/stjude-rust-labs/sprocket/pull/1115)).
+* A cache leaf whose sparse-checkout metadata cannot be parsed is now evicted
+  and re-cloned, and that metadata is written through a temporary file and a
+  rename ([#1115](https://github.com/stjude-rust-labs/sprocket/pull/1115)).
+* Materialized-tree limit checks read blob sizes from Git object headers and
+  fail when an object cannot be read, instead of loading each blob's contents
+  and silently skipping unreadable ones
+  ([#1115](https://github.com/stjude-rust-labs/sprocket/pull/1115)).
 
 ## 0.3.1 - 2026-08-05
